Move expander related code to subdirectory
The expander related code has been moved to a subdirectory, so as to be
able to use it as a submodule. Since the task it is supposed to do can
be done very isolated, I deemed it more useful to have it as a module
rather than cluttering the utils file or main directory. Since I do
expect this to become quite cumbersome do get correctly.