Move conv_2_word_pos to expander.utils
The previously protected conv_2_word_pos has been moved to the utilities
folder, since it is a general utility and might find use cases elsewhere
as well. The expander.py should only have expand_contractions as a
public function all the others should be kept protected.

This is in the mindset of letting the modules do one thing, but that
thing is done well.