Add correct handling of double contraction
Previously contractions like "Who'd've" thought were not dealt with
correctly. This has now been fixed by including a check into
conv_2_word_pos that makes sure that no apostrophe is hidden inside a
word. Note that this makes the assumption that apostrophes in the middle
of the word (i.e. not the end or the beginning) are contractions.