Refactor the expander code for further extensions
The expander code has been completely refactored to make it more
modularizable. Beforehand, everything was handled in a single pair of
for-loops which have now been replaced by multiple sub-routines and more
comprehensive handling of the replacement.

This structure will allow for easier handling of the ambiguous cases.

For now previous output is reproduced if correct and ambiguous output
marked as such and left unchanged.