Generalize the load_stanford_pos function
load_stanford_pos has been generalized and renamed to load_stanford. The
model flag can be used now to either load the pos tagger as before or
the named-entity recognition model of stanford. This is done to finally
be able to replace names and the likes in text with generic lexical
items like 'it' to potentially increase the accuracy of the POS-tagger.

Furthermore, the README has been updated.