Change the docstring formatting
The docstrings now name the args and return information before the
general information about the function. This is done since if one calls
the function (in vim) one gets a preview of the first few lines of the
function. Usually one wants to know what form the arguments have to have
rather than the general information. Since if one calls the function one
likely knows what it's purpose is.