Implement NER-functionality, handle n't case
It is now possible using the use_ner flag on disambiguate and expander
to replaced named entities with a tag (usually <NE>). Which, upon
POS-tagging, will be replaced with it. This should increase accuracy for
named entities. Since usually their pos-tags are not exactly accurate.
This comes at a performance cost though (obviously). The dictionary
contractions.yaml has been updated accordingly to include the cases with
<NE>.

The negation case "n't" has been now correctly implemented as a special
case. Since the word_tokenizer already splits off n't splitting at
apostrophes was not the way to go. It was thus added as a special case
and is now correctly recognized.

Smaller bugfixes and updates of the README have also happened.