Fix bugs, refactor and add features
Smaller bugs have been fixed that were revealed by the new test cases in
expander.py.

The code of expander.py has been refactored to have perfect pylint
score. This basically entailed splitting up long and hard-to-understand
functions into smaller chunks.

The feature has been added which allows replacing contractions with
arbitrary length expansions as opposed to both having to have the same
length. That means that expansions like Ma'am (which has length two, ma
and 'am) to Madam are now possible.

Furthermore, the cases of 'tis and 'twas were added to the dictionary,
though this might be dangerous as it is just implemented as 't
to it due to the way the tokenization happens on the longer
version.

Lastly, add_ne_tag adds the occurence values relatively now, and counts
the ones from 'it' only half (since most named entities are persons),
this is still a bit arbitrary.